haystack/whoosh/requirements.txt
Django==4.1
django-haystack==3.2.1
djangorestframework==3.13.1
Whoosh==2.7.4